+++ CHANGELOG.md
@@ -5,11 +5,30 @@ All notable changes to the "archetype-language-extension" extension will be docu
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.17.0] - 2020-01-17
+### Removed
+ - Remove Michelson language support
+ - Add command : Generate Michelson Storage via LIGO
 
-- Syntax highlighting
-- Snippets
-- Icons
+## [0.13.0] - 2019-10-24
+### Changed
+ - Update syntax highlightning for archetype
 
-## [0.0.1] - 2019-??-??
+## [0.12.0] - 2019-10-17
+### Changed
+ - Improve syntax highlightning for archetype
+
+## [0.11.0] - 2019-10-15
 ### Added
+ - Archetype properties explorer view
+
+### Changed
+ - Update syntax highlightning for archetype
+
+## [0.8.0] - 2019-09-18
+### Added
+ - Inital release
+ - Syntax highlighting for archetype and michelson language
+ - Snippets
+ - Icons
+ - LSP bindings